fix(platform): scope MiniMax usage accounting patch
Keep the MiniMax reasoning-token usage backport for chat completions, but bind the request-path accounting wrapper only to MiniMax chat-serving instances. This avoids routing other reasoning models through the MiniMax usage path and keeps unrelated parser failures out of MiniMax patch frames.

Remove the generic chat usage patch and cover the MiniMax, wrapped-parser, non-MiniMax, and GLM-composition-sensitive paths with focused unit tests.

(cherry picked from commit ec739c18270b0e8a099c05c5ba1a157c9f2490be)
